pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Autonomous web research agent with local/cloud LLM options"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search<7.0.0",
    "feedparser>=6.0.12",
    "langchain>=0.2.10",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.2.10",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.1.10",
    "wikipedia>=1.4.0",
    "gradio>=5.0.0",
    "google-api-python-client>=2.190.0",
    "pyinstaller>=6.19.0",
    "pystray>=0.19.5",
    "pillow>=12.1.1",
]
